CHG | Refactor app initialization by moving code from app.ts to index.ts for improved structure
